+++ src/Notifier/Adapters/SmsChannelAdapter.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Bow\Messaging\Adapters;
+namespace Bow\Notifier\Adapters;
 
 use Bow\Database\Barry\Model;
-use Bow\Messaging\Contracts\ChannelAdapterInterface;
-use Bow\Messaging\Messaging;
+use Bow\Notifier\Contracts\ChannelAdapterInterface;
+use Bow\Notifier\Notifier;
 use InvalidArgumentException;
 use Twilio\Exceptions\ConfigurationException;
 use Twilio\Rest\Client;
@@ -40,35 +40,35 @@ class SmsChannelAdapter implements ChannelAdapterInterface
     }
 
     /**
-     * Send message via SMS
+     * Send notifier via SMS
      *
      * @param  Model     $context
-     * @param  Messaging $message
+     * @param  Notifier $notifier
      * @return void
      */
-    public function send(Model $context, Messaging $message): void
+    public function send(Model $context, Notifier $notifier): void
     {
-        if (!method_exists($message, 'toSms')) {
+        if (!method_exists($notifier, 'toSms')) {
             return;
         }
 
-        $this->sendWithTwilio($context, $message);
+        $this->sendWithTwilio($context, $notifier);
     }
 
     /**
-     * Send the message via SMS using Twilio
+     * Send the notifier via SMS using Twilio
      *
      * @param  Model     $context
-     * @param  Messaging $message
+     * @param  Notifier $notifier
      * @return void
      */
-    private function sendWithTwilio(Model $context, Messaging $message): void
+    private function sendWithTwilio(Model $context, Notifier $notifier): void
     {
-        $data = $message->toSms($context);
+        $data = $notifier->toSms($context);
 
-        $account_sid = config('messaging.twilio.account_sid');
-        $auth_token = config('messaging.twilio.auth_token');
-        $this->from_number = config('messaging.twilio.from');
+        $account_sid = config('notifier.twilio.account_sid');
+        $auth_token = config('notifier.twilio.auth_token');
+        $this->from_number = config('notifier.twilio.from');
 
         if (!$account_sid || !$auth_token || !$this->from_number) {
             throw new InvalidArgumentException('Twilio credentials are required');
@@ -77,17 +77,14 @@ class SmsChannelAdapter implements ChannelAdapterInterface
         $this->client = new Client($account_sid, $auth_token);
 
         if (!isset($data['to']) || !isset($data['message'])) {
-            throw new InvalidArgumentException('The phone number and message are required');
+            throw new InvalidArgumentException('The phone number and notifier are required');
         }
 
         try {
-            $this->client->messages->create(
-                $data['to'],
-                [
+            $this->client->notifiers->create($data['to'], [
                 'from' => $this->from_number,
-                'body' => $data['message']
-                ]
-            );
+                'body' => $data['notifier']
+            ]);
         } catch (\Exception $e) {
             throw new \RuntimeException('Error while sending SMS: ' . $e->getMessage());
         }